+++ src/Transformer/AeatResponseTransformer.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Flux\VerifactuBundle\Transformer;
 
+use Flux\VerifactuBundle\Contract\AeatResponseInterface;
 use Flux\VerifactuBundle\Dto\AeatResponseDto;
 use josemmo\Verifactu\Models\Responses\AeatResponse;
-use Symfony\Component\Serializer\Exception\ExceptionInterface;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 
 final readonly class AeatResponseTransformer extends BaseTransformer
@@ -14,8 +16,19 @@ final readonly class AeatResponseTransformer extends BaseTransformer
     public const FORMAT_JSON = 'json';
 
     public function __construct(
-        private SerializerInterface $serializer,
+        private DenormalizerInterface&NormalizerInterface&SerializerInterface $serializer,
     ) {
+    }
+
+    public function transformInterfaceToDto(AeatResponseInterface $input): AeatResponseDto
+    {
+        return new AeatResponseDto(
+            csv: $input->getCsv(),
+            submittedAt: $input->getSubmittedAt(),
+            waitSecond: $input->getWaitSeconds(),
+            status: $input->getStatus(),
+            items: $input->getItems(),
+        );
     }
 
     public function transformModelToDto(AeatResponse $model): AeatResponseDto
@@ -29,17 +42,21 @@ final readonly class AeatResponseTransformer extends BaseTransformer
         );
     }
 
-    /**
-     * @throws ExceptionInterface
-     */
+    public function transformDtoToArray(AeatResponseDto $dto): array
+    {
+        return $this->serializer->normalize($dto, self::FORMAT_JSON);
+    }
+
+    public function transformArrayToDto(array $aeatResponseArray): AeatResponseDto
+    {
+        return $this->serializer->denormalize($aeatResponseArray, AeatResponseDto::class);
+    }
+
     public function transformDtoToJson(AeatResponseDto $dto): string
     {
         return $this->serializer->serialize($dto, self::FORMAT_JSON);
     }
 
-    /**
-     * @throws ExceptionInterface
-     */
     public function transformJsonToDto(string $json): AeatResponseDto
     {
         return $this->serializer->deserialize($json, AeatResponseDto::class, self::FORMAT_JSON);